+++ DesignToolbox/DesignToolboxUITests/Design System Toolbox App/Components/Controls/Radio/RadioItemUITests.swift
@@ -31,6 +31,7 @@ final class RadioItemUITests: AppTestCase {
         let app = launchApp()
         goToComponentsSheet(app)
         swipeFromDownToUp(app)
+        swipeFromDownToUp(app)
         tapButton(withWording: "app_components_radioButton_label", app)
         tapButton(withWording: "Radio button item", app)