+++ DesignToolbox/DesignToolboxUITests/Design System Toolbox App/Components/Controls/Radio/RadioPickerUITests.swift
@@ -37,6 +37,7 @@ final class RadioPickerUITests: AppTestCase {
         let app = launchApp()
         goToComponentsSheet(app)
         swipeFromDownToUp(app)
+        swipeFromDownToUp(app)
         tapButton(withWording: "app_components_radioButton_label", app)
         tapButton(withWording: "app_components_radioPicker_label", app)